# server/openapi.json
{
  "components" : {
    "schemas" : {
...
      "Announcement" : {
        "properties" : {
          "bodyAr" : {
            "type" : "string"
          },
          "bodyEn" : {
            "type" : "string"
          },
          "classId" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "curriculum" : {
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "grade" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "schoolId" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "teacherId" : {
            "type" : "string"
          },
          "teacherName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreateAnnouncementRequest" : {
        "properties" : {
          "bodyAr" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "bodyEn" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "classId" : {
            "minLength" : 1,
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "bodyEn", "classId" ],
        "type" : "object"
      },
...
      "CreateSchoolRequest" : {
        "properties" : {
          "code" : {
            "pattern" : "^[A-Z0-9]{6}$",
            "type" : "string"
          },
          "curriculumOptions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "gradeOptions" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
...
      "DashboardUser" : {
        "properties" : {
          "createdAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "impersonatedBy" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "lastLoginAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "mustChangePassword" : {
            "type" : "boolean"
          },
          "photoUrl" : {
            "type" : "string"
          },
          "platformName" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "schoolId" : {
            "type" : "string"
          },
          "schoolName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "JoinSchoolInfo" : {
        "properties" : {
          "curriculumOptions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "gradeOptions" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "logoUrl" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "theme" : {
            "$ref" : "#/components/schemas/SchoolTheme"
          }
        },
        "type" : "object"
      },
...
      "School" : {
        "properties" : {
          "children" : {
            "format" : "int32",
            "type" : "integer"
          },
          "classes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "curriculumOptions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "featureFlags" : {
            "type" : "string"
          },
          "gradeOptions" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "id" : {
            "type" : "string"
          },
          "lessons" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "teachers" : {
            "format" : "int32",
            "type" : "integer"
          },
          "theme" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SchoolLogo" : {
        "properties" : {
          "logoUrl" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SchoolLogoRequest" : {
        "properties" : {
          "email" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SchoolSummary" : {
        "properties" : {
          "children" : {
            "format" : "int32",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lessons" : {
            "format" : "int32",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "teachers" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "SchoolTheme" : {
        "properties" : {
          "accent" : {
            "example" : "#CC2A0F",
            "type" : "string"
          },
          "appName" : {
            "maxLength" : 60,
            "minLength" : 0,
            "type" : "string"
          },
          "fontChoice" : {
            "enum" : [ "baloo", "nunito", "fredoka" ],
            "type" : "string"
          },
          "ground" : {
            "example" : "#F3F2F2",
            "type" : "string"
          },
          "logoUrl" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "mascotColor" : {
            "example" : "#598FB8",
            "type" : "string"
          },
          "primary" : {
            "example" : "#FFFFFF",
            "type" : "string"
          },
          "primaryInk" : {
            "example" : "#201E1D",
            "type" : "string"
          },
          "softBorder" : {
            "example" : "#D9D6D2",
            "type" : "string"
          },
          "worldPalettes" : {
            "additionalProperties" : {
              "$ref" : "#/components/schemas/WorldPalette"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "SchoolWizardRequest" : {
        "properties" : {
          "flags" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : "object"
          },
          "manager" : {
            "$ref" : "#/components/schemas/WizardManagerInput"
          },
          "school" : {
            "$ref" : "#/components/schemas/CreateSchoolRequest"
          },
          "theme" : {
            "$ref" : "#/components/schemas/SchoolTheme"
          }
        },
        "required" : [ "manager", "school" ],
        "type" : "object"
      },
      "SchoolWizardResponse" : {
        "properties" : {
          "flags" : {
            "additionalProperties" : {
              "type" : "boolean"
            },
            "type" : "object"
          },
          "invited" : {
            "type" : "boolean"
          },
          "school" : {
            "$ref" : "#/components/schemas/School"
          },
          "theme" : {
            "$ref" : "#/components/schemas/SchoolTheme"
          },
          "user" : {
            "$ref" : "#/components/schemas/DashboardUser"
          }
        },
        "type" : "object"
      },
...
      "UpdateSchoolRequest" : {
        "properties" : {
          "curriculumOptions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "gradeOptions" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WizardManagerInput" : {
        "properties" : {
          "displayName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        },
        "required" : [ "email" ],
        "type" : "object"
      },
      "WorldPalette" : {
        "properties" : {
          "deep" : {
            "type" : "string"
          },
          "ink" : {
            "type" : "string"
          },
          "primary" : {
            "type" : "string"
          },
          "soft" : {
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/admin/schools" : {
      "get" : {
        "operationId" : "listSchools",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/SchoolSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      },
      "post" : {
        "operationId" : "createSchool",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSchoolRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/School"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      }
    },
    "/admin/schools/wizard" : {
      "post" : {
        "operationId" : "createSchoolWithWizard",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SchoolWizardRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchoolWizardResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "Schools" ]
      }
    },
    "/admin/schools/{id}" : {
      "get" : {
        "operationId" : "school",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/School"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      },
      "patch" : {
        "operationId" : "updateSchool",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateSchoolRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/School"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      }
    },
...
    "/children/{id}/announcements" : {
      "get" : {
        "operationId" : "childAnnouncements",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Announcements" ]
      }
    },
...
    "/schools/by-code/{code}" : {
      "get" : {
        "operationId" : "schoolByCode",
        "parameters" : [ {
          "in" : "path",
          "name" : "code",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JoinSchoolInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      }
    },
    "/schools/logo" : {
      "post" : {
        "operationId" : "schoolLogo",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SchoolLogoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchoolLogo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Schools" ]
      }
    },
...
    "/teacher/announcements" : {
      "get" : {
        "operationId" : "teacherAnnouncements",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Announcement"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Announcements" ]
      },
      "post" : {
        "operationId" : "createAnnouncement",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAnnouncementRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Announcement"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "Announcements" ]
      }
    },
    "/teacher/announcements/{id}" : {
      "delete" : {
        "operationId" : "deleteAnnouncement",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "Announcements" ]
      }
    },
...
  },
...
}